Add files via upload
This report provides an analytics methodology and insights about Real Estate brokerage market. First, we develop our methodology that integrates various statistical and analytical methods to construct and interpret the graphs in the analysis part. Then, the applicability of the methodology is illustrated with a case study, where data for top 50 real estate brokerage firms in Dubai are analyzed. In particular, graphs visualizations provide a better understanding of the real estate broker, as demonstrated in the case study.